+++ tmetadata/src/main/resources/application.conf
@@ -14,29 +14,3 @@ service {
 
 environment = "Dev"
 environment = ${?APPLICATION_ENVIRONMENT}
-
-
-
-akka {
-  # Loggers to register at boot time (akka.event.Logging$DefaultLogger logs
-  # to STDOUT)
-  loggers = ["akka.event.slf4j.Slf4jLogger"]
-
-  # Log level used by the configured loggers (see "loggers") as soon
-  # as they have been started; before that, see "stdout-loglevel"
-  # Options: OFF, ERROR, WARNING, INFO, DEBUG
-  loglevel = "ERROR"
-
-  # Log level for the very basic logger activated during ActorSystem startup.
-  # This logger prints the log messages to stdout (System.out).
-  # Options: OFF, ERROR, WARNING, INFO, DEBUG
-  //stdout-loglevel = "DEBUG"
-
-  # Filter of log events that is used by the LoggingAdapter before
-  # publishToKafka log events to the eventStream.
-  logging-filter = "akka.event.DefaultLoggingFilter"
-
-  #Dead-letter logging settings
-  #log-dead-letters = off
-  #log-dead-letters-during-shutdown = off
-}
