--- tmetadata/src/main/resources/logback.xml
@@ -1,9 +1,0 @@
-<configuration>
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder> <pattern>[%highlight(%-5level)] %cyan(%logger{5}): %msg %n</pattern> </encoder>
-    </appender> <logger name="application" level="INFO" />
-    <logger name="akka" level="INFO" />
-    <root level="INFO">
-        <appender-ref ref="STDOUT" />
-    </root>
-</configuration>
