+++ tmetadata/src/main/scala/no/sysco/middleware/kafka/tmetadata/Main.scala
@@ -5,26 +5,29 @@ import akka.event.Logging
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.Http.ServerBinding
 import akka.stream.ActorMaterializer
-import no.sysco.middleware.kafka.tmetadata.actors.KafkaService
-import no.sysco.middleware.kafka.tmetadata.rest.HttpService
+import no.sysco.middleware.kafka.tmetadata.application.KafkaTopicsMetadataService
+import no.sysco.middleware.kafka.tmetadata.rest.HttpRoutes
+
 import scala.concurrent.duration._
 import scala.concurrent.{Await, ExecutionContextExecutor, Future}
 
-// todo: add supervisor and extends it
-object Main extends App with HttpService {
+
+object Main extends App with HttpRoutes {
 
   implicit val system = ActorSystem("akka-reactive-kafka-app")
   implicit val materializer: ActorMaterializer = ActorMaterializer()
   implicit val executionContext: ExecutionContextExecutor = system.dispatcher
 
   val config = Config.loadConfig()
-  val log = Logging(system, this.getClass.getName)
+//  val log = Logging(system, this.getClass.getName)
+
+  override def kafkaService = new KafkaTopicsMetadataService(config)
 
   //Start the akka-http server and listen for http requests
   val akkaHttpServer = startAkkaHTTPServer(config.rest.host, config.rest.port)
 
   // create producer
-  override def kafkaService = system.actorOf(KafkaService.props(config), "kafka-service-actor")
+
 
   private def startAkkaHTTPServer(host: String, port: Int): Future[ServerBinding] = {
     println(s"Waiting for http requests at http://$host:$port/")
