+++ tmetadata/src/main/scala/no/sysco/middleware/kafka/tmetadata/rest/HttpRoutes.scala
@@ -3,6 +3,6 @@ package no.sysco.middleware.kafka.tmetadata.rest
 import no.sysco.middleware.kafka.tmetadata.rest.routes.{AppRoutes, HealthRoutes, MetricRoutes}
 import akka.http.scaladsl.server.Directives._
 
-trait HttpService extends AppRoutes with MetricRoutes with HealthRoutes {
+trait HttpRoutes extends AppRoutes with MetricRoutes with HealthRoutes {
   def routes = appHttpRoutes ~ healthRoutes ~ metricRoutes
 }
